Refactor `XRPLNetwork` to `Common` (#286)
* skeleton of common

* export enum to common
