+++ src/Common/xrpl-network.ts
@@ -1,0 +1,10 @@
+/**
+ * Possible networks to resolve
+ */
+enum XRPLNetwork {
+  Dev = 'devnet',
+  Test = 'testnet',
+  Main = 'mainnet',
+}
+
+export default XRPLNetwork
